# docs/channel/volc-passthrough/apifox-seedance2.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "seedance-2官方透传",
    "version": "1.0.0",
...
    "x-apifox-folder": "seedance-2官方透传"
  },
  "servers": [
    {
      "url": "{{baseUrl}}",
      "description": "站点地址。在 Apifox 环境变量里把 baseUrl 设为 https://your-site.com（不带尾斜杠）"
    }
  ],
...
  "paths": {
...
    "/volc/?Action=CreateAsset&Version=2024-01-01": {
      "post": {
        "tags": [
          "二、控制面 · 12 个 Action（逐个）"
        ],
        "summary": "3.1 CreateAsset 上传素材",
        "operationId": "CreateAsset",
        "description": "上传一张图/一段音视频到指定组，返回 `asset-` 开头的 id。\n\n| 字段 | 类型 | 必选 | 说明 |\n|---|---|---|---|\n| `AssetType` | string | 是 | `Image` ｜ `Video` ｜ `Audio` |\n| `URL` | string | 是 | 素材的**公网可访问地址** |\n| `GroupId` | string | 是 | 所属素材组 id（2.1 返回的） |\n| `Name` | string | 是 | ≤64 字符 |\n\n**只接受 URL，不接受 Base64。** 站点不代传文件，URL 必须是火山侧能直接下载的地址。\n\n`Name` **只用于 3.2 的模糊搜索，不会进入模型推理**。想让画面里出现什么，写在视频提交的提示词里，不要指望素材名生效。\n\n`GroupId` 必须是你自己的组，否则 404 `resource_not_found`。**站点在这一刻把 `user_id + channel_id + asset_id` 记进账本。**\n\n**规格限制**：\n\n| | 图像 | 视频 | 音频 |\n|---|---|---|---|\n| 格式 | jpeg png webp bmp tiff gif heic heif | mp4 mov | wav mp3 |\n| 大小 | < 30 MB | ≤ 200 MB | ≤ 15 MB |\n| 时长 | — | [2, 30] s | [2, 30] s |\n| 宽高比 | (0.4, 2.5) | [0.4, 2.5] | — |\n| 边长 px | (300, 6000) | [300, 6000] | — |\n| 总像素 | — | [407696, 8295044] | — |\n| 帧率 | — | [24, 60] FPS | — |\n| 分辨率 | — | 480p / 720p / 1080p / 4K | — |\n\n上传后素材进入 `Processing`，转码完成才变 `Active`。**`Processing` 状态不能用 `asset://` 引用**，需先按 3.3 轮询到 `Active`。",
        "security": [
          {
            "siteToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "上传图片": {
                  "summary": "最常用。URL 必须是公网可访问的",
                  "value": "{\n  \"AssetType\": \"Image\",\n  \"URL\": \"https://example.com/hero-face.png\",\n  \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n  \"Name\": \"主角正脸\"\n}"
                },
                "上传视频": {
                  "summary": "时长必须在 [2,30] s、≤200 MB、帧率 [24,60] FPS",
                  "value": "{\n  \"AssetType\": \"Video\",\n  \"URL\": \"https://example.com/hero-clip.mp4\",\n  \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n  \"Name\": \"主角动作参考\"\n}"
                },
                "上传音频": {
                  "summary": "wav/mp3，≤15 MB，时长 [2,30] s",
                  "value": "{\n  \"AssetType\": \"Audio\",\n  \"URL\": \"https://example.com/voice.wav\",\n  \"GroupId\": \"assetgroup-20260410113901-qk2mv\",\n  \"Name\": \"主角音色\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功。HTTP 200，信封为 {ResponseMetadata, Result}",
            "content": {
              "application/json": {
                "examples": {
                  "成功": {
                    "summary": "写进集合变量 asset_id；此时 Status 还是 Processing",
                    "value": "{\n  \"ResponseMetadata\": {\n    \"RequestId\": \"202609140804468423AF90FD6CA6BA26F6\",\n    \"Action\": \"CreateAsset\",\n    \"Version\": \"2024-01-01\",\n    \"Service\": \"ark\",\n    \"Region\": \"cn-beijing\"\n  },\n  \"Result\": {\n    \"Id\": \"asset-20260410114236-8cdfz\"\n  }\n}"
                  },
                  "【反向】GroupId 不是自己的": {
                    "summary": "归属校验在发出上游请求之前就被站点拦掉",
                    "value": "{\n  \"ResponseMetadata\": {\n    \"RequestId\": \"202609140804468423AF90FD6CA6BA26F6\",\n    \"Action\": \"CreateAsset\",\n    \"Version\": \"2024-01-01\",\n    \"Region\": \"cn-beijing\",\n    \"Error\": {\n      \"CodeN\": 404,\n      \"Code\": \"resource_not_found\",\n      \"Message\": \"素材组不存在或不属于当前账号: assetgroup-20260410999999-other\"\n    }\n  }\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "siteToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "站点令牌，形如 sk-xxxxxxxx。数据面与控制面透传都用它：Authorization: Bearer {{token}}。在 Apifox 环境变量里配置 token，并把本方案设为全局认证。真值请填自己站点的令牌，不要用示例里的占位符。"
      },
      "siteCookie": {
        "type": "apiKey",
        "in": "header",
        "name": "Cookie",
        "description": "站点登录态，形如 session=xxxxx。仅 /api/volc_asset/* 这五个站点内部接口用；也可改用站点 Access Token 走同名头部。"
      }
    },
...
  },
  "security": [
    {
      "siteToken": []
    }
  ]
}